refactor: api 연결 리팩토링
